fix: HikCam 버그수정완료, exposureTime, Gain 일부 수정

+++ JidamVision/Grab/WebCam.cs
@@ -136,7 +136,7 @@ namespace JidamVision.Grab
             exposure = (long)_capture.Get(VideoCaptureProperties.Exposure);
             return true;
         }
-        internal override bool SetGain(long gain)
+        internal override bool SetGain(float gain)
         {
             if (_capture == null)
                 return false;
@@ -144,7 +144,7 @@ namespace JidamVision.Grab
             _capture.Set(VideoCaptureProperties.Gain, gain);
             return true;
         }
-        internal override bool GetGain(out long gain)
+        internal override bool GetGain(out float gain)
         {
             gain = 0;
             if (_capture == null)
